Recommitting code over from Jupyter to VScode
